requirements.txt
faiss-cpu
pandas
numpy
sentence-transformers
gpt4all
python-dotenv